requirements.txt
streamlit>=1.30.0
google-genai
python-dotenv>=1.0.0
